Fix HW 9

+++ src/s2-homeworks/hw09/Clock.tsx
@@ -38,8 +38,8 @@ function Clock() {
     const stringDate = new Intl.DateTimeFormat('ru').format(date) || <br/> // день.месяц.год (01.02.2022) // пишут студенты, варианты 01.02.0123/01.02.-123/01.02.12345 не рассматриваем
 
     // день недели на английском, месяц на английском (https://learn.javascript.ru/intl#intl-datetimeformat)
-    const stringDay = new Intl.DateTimeFormat('en-US', {weekday: 'short'}).format(date) || <br/> // пишут студенты
-    const stringMonth = new Intl.DateTimeFormat('en-US', {month: 'short'}).format(date) || <br/> // пишут студенты
+    const stringDay = new Intl.DateTimeFormat('en-US', {weekday: 'long'}).format(date) || <br/> // пишут студенты
+    const stringMonth = new Intl.DateTimeFormat('en-US', {month: 'long'}).format(date) || <br/> // пишут студенты
 
     return (
         <div className={s.clock}>
